with to and from with chatgpt after a subscription
with to and from with chatgpt 4 after a subscription, this seems promising as a draft of a program to sync
top level media folders across disks. 
some top level media folders may be duplicated on another 1 or more disks and if so they will be syn'd from the underlying file system of 
the 'first' top level media folder, ffd, the ffd being defined as the first matching disk containing that top level media folder
when iterated left to right of underlying disks specified in the mergerfs mount line fstab . decypher that if you can.